Updated type casts for size_t values.

+++ test/test_sign_async.c
@@ -203,7 +203,7 @@ int main(int argc, char **argv) {
 
 	{
 		size_t count = atoi(argv[ARGV_MAX_REQUEST_COUNT]);
-		KSI_LOG_info(ksi, "Setting max request count to: %u", (unsigned)count);
+		KSI_LOG_info(ksi, "Setting max request count to: %llu", (unsigned long long)count);
 		if (count) {
 			res = KSI_AsyncService_setOption(as, KSI_ASYNC_OPT_MAX_REQUEST_COUNT, (void*)count);
 			if (res != KSI_OK) {
@@ -215,7 +215,7 @@ int main(int argc, char **argv) {
 
 	{
 		size_t size = atoi(argv[ARGV_REQUEST_CACHE_SIZE]);
-		KSI_LOG_info(ksi, "Setting request cache size to: %u", (unsigned)size);
+		KSI_LOG_info(ksi, "Setting request cache size to: %llu", (unsigned long long)size);
 		if (size) {
 			res = KSI_AsyncService_setOption(as, KSI_ASYNC_OPT_REQUEST_CACHE_SIZE, (void*)size);
 			if (res != KSI_OK) {
@@ -227,7 +227,7 @@ int main(int argc, char **argv) {
 	}
 
 	nof_requests = atoi(argv[AGRV_NOF_TEST_REQUESTS]);
-	KSI_LOG_info(ksi, "Nof test requests: %lu", (unsigned long)nof_requests);
+	KSI_LOG_info(ksi, "Nof test requests: %llu", (unsigned long long)nof_requests);
 	do {
 		size_t received = 0;
 
@@ -235,7 +235,7 @@ int main(int argc, char **argv) {
 			if (reqHandle == NULL) {
 				KSI_DataHash *hshRef = NULL;
 
-				KSI_LOG_info(ksi, "Request #: %lu", (unsigned long)req_no);
+				KSI_LOG_info(ksi, "Request #: %llu", (unsigned long long)req_no);
 
 				/* Get the hash value of the input file. */
 				res = createHash(ksi, req_no, &reqHsh);
@@ -436,8 +436,8 @@ int main(int argc, char **argv) {
 
 	res = KSI_OK;
 cleanup:
-	printf("Succeeded request: %lu.\n", (unsigned long)succeeded);
-	printf("Failed request   : %lu.\n", (unsigned long)(nof_requests - succeeded));
+	printf("Succeeded request: %llu.\n", (unsigned long long)succeeded);
+	printf("Failed request   : %llu.\n", (unsigned long long)(nof_requests - succeeded));
 	printf("Spent time (sec) : %.0f.\n", difftime(time(NULL), start));
 
 	if (res != KSI_OK && ksi != NULL) {
